backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
supabase>=2.28.0
openai>=1.50.0
httpx>=0.27.0
python-multipart>=0.0.12
python-dotenv>=1.0.0
python-docx>=1.1.2
pymupdf>=1.25.0
pydantic-settings>=2.5.0
ollama>=0.3.0
json-repair>=0.35.0
